% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.iris.R
\name{plot.iris}
\alias{plot.iris}
\title{plot.iris}
\usage{
\method{plot}{iris}(x, labels,
  main = "A UMAP visualization of the Iris dataset",
  colors = c("#ff7f00", "#e377c2", "#17becf"), pad = 0.1, cex = 0.65,
  pch = 19, add = FALSE, legend.suffix = "", cex.main = 1,
  cex.legend = 1)
}
\value{
A character vector
}
\description{
A function to plot umap figures with iris data.
}
\examples{
plot.iris()
}
